You can connect to server on same wifi. See details below.
If SQL is disabled, you can connect to the server using an ip and port.
You also need to provide a temporary name.  You'll be taken to the lobby
similar to online play and you can see people that are on the same lobby
as you.  It's just that you won't be able to see your online friends.